Fixed logic surrounding original vs fp9 handshake detection
Evostream Media Server sends back a version of 0 but is still using the fp9 digest handshake.  The original code was mistaking this as an originally specced RTMP handshake, but that is not totally accurate because it was still storing digests.  

So now it only checks the version field as zero if no digest could be found.