When logging io more reliably separate out by rtmp bytes from handshake.
Every once and a while we will receive a single TCP packet that contains both the end part of the handshake and the start of the actual rtmp messages, making using debugging tools difficult.  We now are finding any bytes remaining from processing a handshake and logging all non-handshake bytes that come down the pipe.

Also organized the log files into its own data structure to make it easy to (temporarily or permanently) add more log files that are needed for debugging scenarios.