Allow benchmarks to pass in the number of iterations to process.
For some reason WSL and Linux VMs both are extremely slow running the benchmark compared to bare metal windows.  Not totally sure why that is (if it's a vm or wsl issue) but this allows me to run benchmarks on both while tweaking the iteration count to something reasonable for each.