type 3 chunk may have extended timestamp (#15)
A type 3 chunk followed a type 0 chunk which has extended timestamp may has the extended timestamp too.

some rtmp clients send that kind of type3 chunk, for examples, FMLE, ffmpeg, wechat miniprogram(live-pusher), but some not, like rtmpdump.

nignx-rtmp and lal(q191201771/lal, a mediaserver written by golang) always read extended timestamp of a type 3 chunk in this situation. srs(simple rtmp server) compatible with this situation too.

So I created the pull request, it's logic:

if previous chunk of a type 3 chunk was type 0 and it has extended timestamp, then we read 4 bytes as extended timestamp of the type 3 chunk, compare with the previous chunk timestamp, if they are same, then we assume it has extended timestamp, otherwise it does not have.