Expose `PublishMode` enum
When an RTMP client requests publishing it states what type of publishing it wants to perform (video without recording, video with recording, or video to be appended to an existing file).  While the requested publish mode was passed in with a publish request, the enum itself wasn't publicly exposed so consumers couldn't do anything with it (except print it).