threaded_rtmp_server: signal stream-end to players (#34)
When a client stops publishing a stream, signal this to any
watching clients using ServerSession::finish_playing and then
disconnect the watching clients.

Also make sure that we shut down the write-part of the socket from
Connection's writer thread when dropping a client. Not doing this,
i.e. relying on TcpStream::drop to shutdown the socket, seems
risky as there can be a reference to the socket in the reader
thread, in which case we'd be relying on the client to disconnect.

This change allows a client like FFmpeg to properly detect the
end of a stream. Previously, consuming a stream with the command
'ffmpeg -i rtmp://localhost:1935/a/b out.mp4' and ending the
stream from the publisher's end, the FFmpeg command would 'hang'
and not finish until the rml server is shut down. Now, the FFmpeg
command finishes when the stream ends and produces a valid
output file.

Co-authored-by: Jonathan Murray <jonathanmurray@spotify.com>