Fix tests (#36)
* Fix test for window ack size

* Do not preallocate `VecDeque` with capacity

We removed setting the size of the chunk size in of the previous
patches. Thus it does not make sense to preallocate it to a specific
number.

* Fix test for checking chunk size after session creation

In previous patches, we introduced a breaking change due to moving
the message for setting the chunk size. Now the chunk size is set
after a successful connection.